Update tcp_client.cpp
For a reason I don't really understand, the msg buffer doesn't seem to get empty at the end of the loop. (The end of a long message can be read on the next shorter message received from the same socket). The memset function allows to empty the buffer before reusing it.