fixed deadlock when using tcp-client close() func
in the tcp-client example, the client close() func was called from
within the client subscriber callback function. this caused a deadlock
when the server disconnected, since the disconnection CB triggered a
close() call.
The redundant call was removed.
Instead, the client will be closed automatically when the object is
distructed, or when the user invoke it from its own thread context.